Rename and make new package
