+++ ast/src/main/java/com/graphicsfuzz/common/ast/stmt/ReturnStmt.java
@@ -25,11 +25,13 @@ public class ReturnStmt extends Stmt {
   private Expr expr;
 
   public ReturnStmt(Expr expr) {
+    assert expr != null;
     this.expr = expr;
   }
 
   public ReturnStmt() {
     this(null);
+    this.expr = null;
   }
 
   public Expr getExpr() {
