+++ generator/src/main/java/com/graphicsfuzz/generator/knownvaluegeneration/VariableFact.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.graphicsfuzz.generator.mutateapi;
+package com.graphicsfuzz.generator.knownvaluegeneration;
 
 public abstract class VariableFact {
   private final Value value;
@@ -35,7 +35,7 @@ public abstract class VariableFact {
 
   /**
    * @return a variable name of this variable fact. This is used by the generator
-   * when it is generating a new variable identifier expression.
+   *     when it is generating a new variable identifier expression.
    */
   public abstract String getVariableName();
 }
